requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
dj-database-url==0.5.0
Django==3.0.8
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.0
flake8==3.8.3
future==0.18.2
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
importlib-metadata==1.7.0
kiwisolver==1.2.0
matplotlib==3.3.0
mccabe==0.6.1
numpy==1.19.1
oauth2==1.9.0.post1
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-openid==2.2.5
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
torch==1.6.0
torchvision==0.7.0
urllib3==1.25.9
whitenoise==5.1.0
zipp==3.1.0

